1. Button in a weired place, 2. Page should not be scrollable, 3. Upgrade plan banner issue redesigned
