+++ airbyte-webapp/src/pages/ConnectionPage/pages/AllConnectionsPage/components/UpgradePlanBar.tsx
@@ -1,28 +1,33 @@
-import React from "react";
+import React, { useState } from "react";
 import { FormattedMessage } from "react-intl";
 import styled from "styled-components";
 
 import { Button } from "components";
 
+import { useUser } from "core/AuthContext";
+import { getRoleAgainstRoleNumber, ROLES } from "core/Constants/roles";
+import { getStatusAgainstStatusNumber, STATUSES } from "core/Constants/statuses";
+import useRouter from "hooks/useRouter";
+import { UnauthorizedModal } from "pages/ConnectionPage/pages/AllConnectionsPage/components/UnauthorizedModal";
+import { RoutePaths } from "pages/routePaths";
+import { SettingsRoute } from "pages/SettingsPage/SettingsPage";
 import { useUserPlanDetail } from "services/payments/PaymentsService";
-
-interface IProps {
-  onUpgradePlan?: () => void;
-}
 
 const Container = styled.div`
   width: 100%;
-  height: 90px;
+  height: 40px;
   background-color: #9596a4;
   display: flex;
   flex-direction: row;
   align-items: center;
   justify-content: center;
+  position: fixed;
+  z-index: 10000;
 `;
 
 const Text = styled.div`
   font-weight: 500;
-  font-size: 16px;
+  font-size: 13px;
   line-height: 24px;
   display: flex;
   align-items: center;
@@ -30,7 +35,9 @@ const Text = styled.div`
   margin-right: 50px;
 `;
 
-export const UpgradePlanBar: React.FC<IProps> = ({ onUpgradePlan }) => {
+export const UpgradePlanBar: React.FC = () => {
+  const { user } = useUser();
+  const { pathname, push } = useRouter();
   const userPlanDetail = useUserPlanDetail();
   const { expiresTime } = userPlanDetail;
 
@@ -41,17 +48,49 @@ export const UpgradePlanBar: React.FC<IProps> = ({ onUpgradePlan }) => {
     const diffDays = Math.ceil(diff / (1000 * 60 * 60 * 24));
     return diffDays;
   };
-  return (
-    <Container>
-      <Text>
-        <FormattedMessage
-          id={remainingDaysForFreeTrial() >= 0 ? "upgrade.plan.trialPeriod.countdown" : "upgrade.plan.trialPeriod.end"}
-          values={{ count: remainingDaysForFreeTrial() }}
-        />
-      </Text>
-      <Button size="lg" black onClick={() => onUpgradePlan?.()}>
-        <FormattedMessage id="upgrade.plan.btn" />
-      </Button>
-    </Container>
-  );
+
+  const isUpgradePlanBar = (): boolean => {
+    let showUpgradePlanBar = false;
+    if (getStatusAgainstStatusNumber(user.status) === STATUSES.Free_Trial) {
+      if (!pathname.split("/").includes(RoutePaths.Payment)) {
+        showUpgradePlanBar = true;
+      }
+    }
+    return showUpgradePlanBar;
+  };
+
+  const [isAuthorized, setIsAuthorized] = useState<boolean>(false);
+
+  const onUpgradePlan = () => {
+    if (
+      getRoleAgainstRoleNumber(user.role) === ROLES.Administrator_Owner ||
+      getRoleAgainstRoleNumber(user.role) === ROLES.Administrator
+    ) {
+      push(`/${RoutePaths.Settings}/${SettingsRoute.PlanAndBilling}`);
+    } else {
+      setIsAuthorized(true);
+    }
+  };
+
+  if (isUpgradePlanBar()) {
+    return (
+      <>
+        <Container>
+          <Text>
+            <FormattedMessage
+              id={
+                remainingDaysForFreeTrial() >= 0 ? "upgrade.plan.trialPeriod.countdown" : "upgrade.plan.trialPeriod.end"
+              }
+              values={{ count: remainingDaysForFreeTrial() }}
+            />
+          </Text>
+          <Button size="m" black onClick={() => onUpgradePlan()}>
+            <FormattedMessage id="upgrade.plan.btn" />
+          </Button>
+        </Container>
+        {isAuthorized && <UnauthorizedModal onClose={() => setIsAuthorized(false)} />}
+      </>
+    );
+  }
+  return null;
 };
