+++ airbyte-webapp/src/views/layout/MainView/MainView.tsx
@@ -1,16 +1,11 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "styled-components";
 
 import { LoadingPage } from "components";
 
-import { useUser } from "core/AuthContext";
-import { getRoleAgainstRoleNumber, ROLES } from "core/Constants/roles";
-import { getStatusAgainstStatusNumber, STATUSES } from "core/Constants/statuses";
 import useRouter from "hooks/useRouter";
-import { UnauthorizedModal } from "pages/ConnectionPage/pages/AllConnectionsPage/components/UnauthorizedModal";
 import { UpgradePlanBar } from "pages/ConnectionPage/pages/AllConnectionsPage/components/UpgradePlanBar";
 import { RoutePaths } from "pages/routePaths";
-import { SettingsRoute } from "pages/SettingsPage/SettingsPage";
 import { ResourceNotFoundErrorBoundary } from "views/common/ResorceNotFoundErrorBoundary";
 import { StartOverErrorView } from "views/common/StartOverErrorView";
 import SideBar from "views/layout/SideBar";
@@ -31,35 +26,11 @@ const Content = styled.div`
 `;
 
 const MainView: React.FC = (props) => {
-  const { user } = useUser();
-  const { pathname, push } = useRouter();
+  const { pathname } = useRouter();
 
-  const [isAuthorized, setIsAuthorized] = useState<boolean>(false);
-
-  // TODO: not the propersolution but works for now
+  // TODO: not the proper solution but works for now
   const isSidebar =
     !pathname.split("/").includes(RoutePaths.Payment) && !pathname.split("/").includes(RoutePaths.PaymentError);
-
-  const isUpgradePlanBar = (): boolean => {
-    let showUpgradePlanBar = false;
-    if (getStatusAgainstStatusNumber(user.status) === STATUSES.Free_Trial) {
-      if (!pathname.split("/").includes(RoutePaths.Payment)) {
-        showUpgradePlanBar = true;
-      }
-    }
-    return showUpgradePlanBar;
-  };
-
-  const onUpgradePlan = () => {
-    if (
-      getRoleAgainstRoleNumber(user.role) === ROLES.Administrator_Owner ||
-      getRoleAgainstRoleNumber(user.role) === ROLES.Administrator
-    ) {
-      push(`/${RoutePaths.Settings}/${SettingsRoute.PlanAndBilling}`);
-    } else {
-      setIsAuthorized(true);
-    }
-  };
 
   return (
     <MainContainer>
@@ -67,8 +38,7 @@ const MainView: React.FC = (props) => {
       <Content>
         <ResourceNotFoundErrorBoundary errorComponent={<StartOverErrorView />}>
           <React.Suspense fallback={<LoadingPage />}>
-            {isAuthorized && <UnauthorizedModal onClose={() => setIsAuthorized(false)} />}
-            {isUpgradePlanBar() && <UpgradePlanBar onUpgradePlan={onUpgradePlan} />}
+            <UpgradePlanBar />
             {props.children}
           </React.Suspense>
         </ResourceNotFoundErrorBoundary>
